+++ app/src/main/java/org/simple/clinic/patient/businessid/Identifier.kt
@@ -15,6 +15,7 @@ import org.simple.clinic.patient.businessid.Identifier.IdentifierType.BpPassport
 import org.simple.clinic.patient.businessid.Identifier.IdentifierType.EthiopiaMedicalRecordNumber
 import org.simple.clinic.patient.businessid.Identifier.IdentifierType.IndiaNationalHealthId
 import org.simple.clinic.patient.businessid.Identifier.IdentifierType.SriLankaNationalId
+import org.simple.clinic.patient.businessid.Identifier.IdentifierType.SriLankaPersonalHealthNumber
 import org.simple.clinic.patient.businessid.Identifier.IdentifierType.Unknown
 import org.simple.clinic.util.Unicode
 import org.simple.clinic.util.room.SafeEnumTypeAdapter
@@ -51,6 +52,7 @@ data class Identifier(
         "$prefix${Unicode.nonBreakingSpace}$subString1${Unicode.nonBreakingSpace}$subString2${Unicode.nonBreakingSpace}$suffix"
       }
       SriLankaNationalId -> value
+      SriLankaPersonalHealthNumber -> value
       is Unknown -> value
     }
   }
@@ -62,6 +64,7 @@ data class Identifier(
       EthiopiaMedicalRecordNumber -> resources.getString(R.string.identifiertype_ethiopia_medical_record_number)
       IndiaNationalHealthId -> resources.getString(R.string.identifiertype_india_national_health_id)
       SriLankaNationalId -> resources.getString(R.string.identifiertype_sri_lanka_national_id)
+      SriLankaPersonalHealthNumber -> resources.getString(R.string.identifiertype_sri_lanka_personal_health_number)
       is Unknown -> resources.getString(R.string.identifiertype_unknown)
     }
   }
@@ -96,6 +99,9 @@ data class Identifier(
     object SriLankaNationalId : IdentifierType()
 
     @Parcelize
+    object SriLankaPersonalHealthNumber : IdentifierType()
+
+    @Parcelize
     data class Unknown(val actual: String) : IdentifierType()
 
     object TypeAdapter : SafeEnumTypeAdapter<IdentifierType>(
@@ -104,7 +110,8 @@ data class Identifier(
             BangladeshNationalId to "bangladesh_national_id",
             EthiopiaMedicalRecordNumber to "ethiopia_medical_record",
             IndiaNationalHealthId to "india_national_health_id",
-            SriLankaNationalId to "sri_lanka_national_id"
+            SriLankaNationalId to "sri_lanka_national_id",
+            SriLankaPersonalHealthNumber to "sri_lanka_personal_health_number"
         ),
         unknownStringToEnumConverter = { Unknown(it) },
         unknownEnumToStringConverter = { (it as Unknown).actual }