+++ app/src/main/java/org/simple/clinic/widgets/PatientSearchResultItemView.kt
@@ -16,12 +16,7 @@ import org.simple.clinic.patient.PatientAgeDetails
 import org.simple.clinic.patient.PatientAgeDetails.Type.EXACT
 import org.simple.clinic.patient.PatientSearchResult
 import org.simple.clinic.patient.businessid.Identifier
-import org.simple.clinic.patient.businessid.Identifier.IdentifierType.BangladeshNationalId
 import org.simple.clinic.patient.businessid.Identifier.IdentifierType.BpPassport
-import org.simple.clinic.patient.businessid.Identifier.IdentifierType.EthiopiaMedicalRecordNumber
-import org.simple.clinic.patient.businessid.Identifier.IdentifierType.IndiaNationalHealthId
-import org.simple.clinic.patient.businessid.Identifier.IdentifierType.SriLankaNationalId
-import org.simple.clinic.patient.businessid.Identifier.IdentifierType.Unknown
 import org.simple.clinic.patient.displayIconRes
 import org.simple.clinic.patient.displayLetterRes
 import org.simple.clinic.router.util.resolveColor
@@ -112,7 +107,7 @@ class PatientSearchResultItemView(
   private fun getId(identifier: Identifier, searchQuery: String): Id {
     val id = when (identifier.type) {
       BpPassport -> BpPassport.shortCode(identifier)
-      BangladeshNationalId, EthiopiaMedicalRecordNumber, IndiaNationalHealthId, SriLankaNationalId, is Unknown -> identifier.value
+      else -> identifier.value
     }
     val indexOfIdentifier = id.indexOf(searchQuery, ignoreCase = true)
     return if (indexOfIdentifier >= 0) {